Adding support of multiple values for unique key col (#116)

+++ dbt/include/vertica/macros/materializations/models/incremental/merge.sql
@@ -5,14 +5,19 @@
  
   merge into {{ target_relation }} as DBT_INTERNAL_DEST
   using {{ tmp_relation }} as DBT_INTERNAL_SOURCE
-
+  
+ 
   {#-- Test 1, find the provided merge columns #}
   {% if merge_columns %}
-    on 
-    {% for column in [merge_columns] %}
+    on
+    {% if merge_columns is string %}
+      DBT_INTERNAL_DEST.{{ adapter.quote(merge_columns) }} = DBT_INTERNAL_SOURCE.{{ adapter.quote(merge_columns) }}
+    {% else %}
+    {% for column in merge_columns -%}
       DBT_INTERNAL_DEST.{{ adapter.quote(column) }} = DBT_INTERNAL_SOURCE.{{ adapter.quote(column) }}
       {%- if not loop.last %} AND {% endif %} 
     {%- endfor %}
+    {% endif %}
   {#-- Test 2, use all columns in the destination table #}
   {% else %}
     on
@@ -54,12 +59,29 @@
     {%- set dest_cols_csv = get_quoted_csv(dest_columns | map(attribute="name")) -%}
 
     {% if unique_key %}
+      {% if unique_key is string %}
         delete from {{ target }}
-            where (
-                {{ unique_key }}) in (
-                select ({{ unique_key }})
-                from {{ source }}
-            );
+            where 
+                ({{ unique_key }}) in (
+                  select ({{ unique_key }})
+                  from {{ source }}
+                )
+              
+                
+            ;
+      {% else %}
+        delete from {{ target }}
+            where 
+              {% for column in unique_key -%}
+                ({{ column }}) in (
+                  select ({{ column }})
+                  from {{ source }}
+                )
+                {%- if not loop.last %} AND {% endif %} 
+              {%- endfor %}
+                
+            ;
+      {% endif%}
 
     {% endif %}
